# forum/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planinarski Forum API",
        "description": "API dokumentacija za Planinarski Forum aplikaciju",
        "contact": {
            "email": "admin@planinarskiforum.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista svih postova",
                "description": "Vraća paginiranu listu postova sa mogućnošću filtriranja i sortiranja",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj postova po stranici (5-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Polje po kome se sortira (created_at, replies_count, likes_count, title)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Smer sortiranja (asc ili desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "theme",
                        "in": "query",
                        "description": "Filter po nazivu teme",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "theme_id",
                        "in": "query",
                        "description": "Filter po ID teme",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Filter po ID autora",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Pretraga po imenu autora",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Pretraga po naslovu i sadržaju",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Datum od (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Datum do (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "sort_by": {
                                                    "type": "string"
                                                },
                                                "sort_dir": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{post}/replies": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Odgovori na post",
                "description": "Vraća listu odgovora za određeni post",
                "operationId": "83fc1cb64834b7333831ab299914d501",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID posta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj odgovora po stranici",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "replies": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post nije pronađen"
                    }
                }
            }
        },
        "/api/posts/{post}/likes": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lajkovi na postu",
                "description": "Vraća listu korisnika koji su lajkovali post",
                "operationId": "6d15384651d27be9a77958636de2aa51",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID posta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj lajkova po stranici",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "likes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post nije pronađen"
                    }
                }
            }
        },
...
    },
...
}